@@ -5,12 +5,12 @@ description: Flutter is an open source framework by Google for building beautifu
 homepage: https://github.com/LHHan/base_project
 
 environment:
-  sdk: '>=2.17.5 <3.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   cupertino_icons: ^1.0.2
-  get: 4.6.6
-  logger: ^1.1.0
+  get: ^4.7.2
+  logger: ^2.5.0
   get_storage: ^2.0.3
   flutter_screenutil: ^5.9.3
   flutter_svg: ^2.0.4
@@ -27,7 +27,7 @@ flutter_intl:
   enabled: true
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
+  flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
 
