@@ -1,4 +1,6 @@
+import 'package:base_project_getx/app/core/utils/app_extension.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class WInputForm extends StatelessWidget {
   const WInputForm({
@@ -8,7 +10,8 @@ class WInputForm extends StatelessWidget {
     this.hintText,
     this.errorText,
     this.suffixIcon,
-    this.obscureText = false,
+    this.prefixIcon,
+    this.obscureTextNotifier,
     this.readOnly = false,
     this.onTap,
     this.onChanged,
@@ -16,6 +19,16 @@ class WInputForm extends StatelessWidget {
     this.focusNode,
     this.keyboardType = TextInputType.text,
     this.textInputAction = TextInputAction.next,
+    this.contentPadding,
+    this.margin,
+    this.style,
+    this.labelStyle,
+    this.hintStyle,
+    this.errorTextColor,
+    this.border,
+    this.focusedBorder,
+    this.errorBorder,
+    this.maxLines = 1,
   }) : super(key: key);
 
   const WInputForm.email({
@@ -25,14 +38,25 @@ class WInputForm extends StatelessWidget {
     this.hintText = 'Enter your email',
     this.errorText,
     this.suffixIcon,
-    this.obscureText = false,
-    this.onTap,
+    this.prefixIcon,
+    this.obscureTextNotifier,
     this.readOnly = false,
+    this.onTap,
     this.onChanged,
     this.onSubmitted,
     this.focusNode,
     this.textInputAction = TextInputAction.next,
     this.keyboardType = TextInputType.emailAddress,
+    this.contentPadding,
+    this.margin,
+    this.style,
+    this.labelStyle,
+    this.hintStyle,
+    this.errorTextColor,
+    this.border,
+    this.focusedBorder,
+    this.errorBorder,
+    this.maxLines = 1,
   }) : super(key: key);
 
   const WInputForm.password({
@@ -42,14 +66,25 @@ class WInputForm extends StatelessWidget {
     this.hintText = 'Enter your password',
     this.errorText,
     this.suffixIcon,
-    this.obscureText = true,
-    this.onTap,
+    this.prefixIcon,
+    this.obscureTextNotifier,
     this.readOnly = false,
+    this.onTap,
     this.onChanged,
     this.onSubmitted,
     this.focusNode,
     this.textInputAction = TextInputAction.done,
     this.keyboardType = TextInputType.visiblePassword,
+    this.contentPadding,
+    this.margin,
+    this.style,
+    this.labelStyle,
+    this.hintStyle,
+    this.errorTextColor,
+    this.border,
+    this.focusedBorder,
+    this.errorBorder,
+    this.maxLines = 1,
   }) : super(key: key);
 
   final TextEditingController? controller;
@@ -57,7 +92,8 @@ class WInputForm extends StatelessWidget {
   final String? hintText;
   final String? errorText;
   final Widget? suffixIcon;
-  final bool obscureText;
+  final Widget? prefixIcon;
+  final ValueNotifier<bool>? obscureTextNotifier;
   final bool readOnly;
   final Function()? onTap;
   final Function(String)? onChanged;
@@ -65,26 +101,71 @@ class WInputForm extends StatelessWidget {
   final FocusNode? focusNode;
   final TextInputAction textInputAction;
   final TextInputType keyboardType;
+  final EdgeInsetsGeometry? contentPadding;
+  final EdgeInsetsGeometry? margin;
+  final TextStyle? style;
+  final TextStyle? labelStyle;
+  final TextStyle? hintStyle;
+  final Color? errorTextColor;
+  final InputBorder? border;
+  final InputBorder? focusedBorder;
+  final InputBorder? errorBorder;
+  final int maxLines;
 
   @override
   Widget build(BuildContext context) {
-    return TextFormField(
-      controller: controller,
-      autocorrect: false,
-      enableSuggestions: false,
-      decoration: InputDecoration(
-        labelText: labelText,
-        hintText: hintText,
-        errorText: errorText,
-        border: const OutlineInputBorder(),
-        suffixIcon: suffixIcon,
+    return Container(
+      margin: margin ?? const EdgeInsets.all(8.0),
+      child: ValueListenableBuilder<bool>(
+        valueListenable: obscureTextNotifier ??
+            ValueNotifier<bool>(obscureTextNotifier?.value ?? false),
+        builder: (context, obscureText, child) {
+          return TextFormField(
+            controller: controller,
+            autocorrect: false,
+            enableSuggestions: false,
+            style: style ?? Get.textTheme.tsBody,
+            decoration: InputDecoration(
+              labelText: labelText,
+              labelStyle: labelStyle ?? Get.textTheme.tsSubTitle,
+              hintText: hintText,
+              hintStyle: hintStyle ?? Get.textTheme.tsHint,
+              errorText: errorText,
+              errorStyle: Get.textTheme.tsBody.copyWith(color: errorTextColor),
+              border: border ?? const OutlineInputBorder(),
+              focusedBorder: focusedBorder ?? const OutlineInputBorder(),
+              errorBorder: errorBorder ?? const OutlineInputBorder(),
+              contentPadding: contentPadding ??
+                  const EdgeInsets.symmetric(horizontal: 12, vertical: 16),
+              prefixIcon: prefixIcon,
+              suffixIcon: suffixIcon ??
+                  (obscureTextNotifier == null
+                      ? null
+                      : obscureText
+                          ? IconButton(
+                              icon: const Icon(Icons.visibility_off),
+                              onPressed: () {
+                                obscureTextNotifier?.value = !obscureText;
+                              },
+                            )
+                          : IconButton(
+                              icon: const Icon(Icons.visibility),
+                              onPressed: () {
+                                obscureTextNotifier?.value = !obscureText;
+                              },
+                            )),
+            ),
+            keyboardType: keyboardType,
+            obscureText: obscureText,
+            onChanged: onChanged,
+            onFieldSubmitted: onSubmitted,
+            focusNode: focusNode,
+            textInputAction: textInputAction,
+            readOnly: readOnly,
+            maxLines: maxLines,
+          );
+        },
       ),
-      keyboardType: keyboardType,
-      obscureText: obscureText,
-      onChanged: onChanged,
-      onFieldSubmitted: onSubmitted,
-      focusNode: focusNode,
-      textInputAction: textInputAction,
     );
   }
 }
