@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class PMaterial extends StatelessWidget {
-  const PMaterial({required this.child, Key? key}) : super(key: key);
+  const PMaterial({required this.child, super.key});
 
   final Widget child;
 
